--- src/main/java/com/travelplan/domain/member/repository/custom/MemberRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.travelplan.domain.member.repository.custom;
-
-import com.travelplan.domain.member.domain.Member;
-import com.travelplan.domain.travel.domain.Travel;
-
-public interface MemberRepositoryCustom {
-
-    Integer findMemberId(Travel travel, String email);
-}
